Fix double-mortalization
Kind of embarrassing that this escaped me at the time: We're using the
"SV *" typemap for the return value of the lookup function. That
typemap already takes care of mortalizing the given SV, because that's
the common use case. So remove the additional, explicit mortalization.
For reference, this was causing warnings of the sort:

  Attempt to free unreferenced scalar: SV 0x266d3c0,
    Perl interpreter: 0x248d010 at t/01_basic.t line 22.

Which I've dissected in great detail at:

  http://blog.booking.com/the-dreaded-attempt-to-free-unreferenced-scalar.html